Require 'yes' flag before submitting
Running without the 'yes' flag simply shows you the issue that would be
submitted, and the list of tracks to which it would be submitted.

My usual workflow is to run without the 'yes' flag, make sure the list
of tracks looks sane, and then run with the 'yes' flag.