Post issues to inactive tracks
If a track that has not yet been activated has implemented an exercise
then the issue is relevant.

If the issue is not specific to an exercise (such as documentation
across the board), then it is also relevant to inactive tracks; perhaps
even more so, since tracks often languish in an unstarted/inactive state
for a long time before someone decides to port enough exercises to make
it live, and the starter repository will have lots of old references in
it.